--- a/src/coreclr/System.Private.CoreLib/src/System/Reflection/ConstructorInvoker.CoreCLR.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Reflection/ConstructorInvoker.CoreCLR.cs
@@ -1,18 +1,21 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.CompilerServices;
-
 namespace System.Reflection
 {
-    internal partial class ConstructorInvoker
+    public partial class ConstructorInvoker
     {
-        public InvocationFlags _invocationFlags;
+        private readonly Signature? _signature;
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private unsafe object? InterpretedInvoke(object? obj, IntPtr* arguments)
+        internal unsafe ConstructorInvoker(RuntimeConstructorInfo constructor) : this(constructor, constructor.Signature.Arguments)
         {
-            return RuntimeMethodHandle.InvokeMethod(obj, (void**)arguments, _method.Signature, isConstructor: obj is null)!;
+            _signature = constructor.Signature;
+            _invokeFunc_RefArgs = InterpretedInvoke;
+        }
+
+        private unsafe object? InterpretedInvoke(object? obj, IntPtr* args)
+        {
+            return RuntimeMethodHandle.InvokeMethod(obj, (void**)args, _signature!, isConstructor: obj is null);
         }
     }
 }