--- a/src/coreclr/System.Private.CoreLib/src/System/Reflection/MethodBaseInvoker.CoreCLR.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Reflection/MethodBaseInvoker.CoreCLR.cs
@@ -1,39 +1,38 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics;
 using System.Reflection.Emit;
 
 namespace System.Reflection
 {
-    public partial class MethodInvoker
+    internal partial class MethodBaseInvoker
     {
         private readonly Signature? _signature;
 
-        internal unsafe MethodInvoker(RuntimeMethodInfo method) : this(method, method.Signature.Arguments)
+        internal unsafe MethodBaseInvoker(RuntimeMethodInfo method) : this(method, method.Signature.Arguments)
         {
             _signature = method.Signature;
-            _invokeFunc_RefArgs = InterpretedInvoke_Method;
             _invocationFlags = method.ComputeAndUpdateInvocationFlags();
-        }
-
-        internal unsafe MethodInvoker(DynamicMethod method) : this(method, method.Signature.Arguments)
-        {
-            _signature = method.Signature;
             _invokeFunc_RefArgs = InterpretedInvoke_Method;
         }
 
-        internal unsafe MethodInvoker(RuntimeConstructorInfo constructor) : this(constructor, constructor.Signature.Arguments)
+        internal unsafe MethodBaseInvoker(RuntimeConstructorInfo constructor) : this(constructor, constructor.Signature.Arguments)
         {
             _signature = constructor.Signature;
-            _invokeFunc_RefArgs = InterpretedInvoke_Constructor;
             _invocationFlags = constructor.ComputeAndUpdateInvocationFlags();
+            _invokeFunc_RefArgs = InterpretedInvoke_Constructor;
         }
 
-        private unsafe object? InterpretedInvoke_Method(object? obj, IntPtr* args) =>
-            RuntimeMethodHandle.InvokeMethod(obj, (void**)args, _signature!, isConstructor: false);
+        internal unsafe MethodBaseInvoker(DynamicMethod method, Signature signature) : this(method, signature.Arguments)
+        {
+            _signature = signature;
+            _invokeFunc_RefArgs = InterpretedInvoke_Method;
+        }
 
         private unsafe object? InterpretedInvoke_Constructor(object? obj, IntPtr* args) =>
             RuntimeMethodHandle.InvokeMethod(obj, (void**)args, _signature!, isConstructor: obj is null);
+
+        private unsafe object? InterpretedInvoke_Method(object? obj, IntPtr* args) =>
+            RuntimeMethodHandle.InvokeMethod(obj, (void**)args, _signature!, isConstructor : false);
     }
 }