--- a/src/coreclr/System.Private.CoreLib/src/System/Reflection/RtFieldInfo.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Reflection/RtFieldInfo.cs
@@ -194,18 +194,17 @@ namespace System.Reflection
 
             CheckConsistency(obj);
 
-            ParameterCopyBackAction _ref = default;
             RuntimeType fieldType = (RuntimeType)FieldType;
             if (value is null)
             {
                 if (RuntimeTypeHandle.IsValueType(fieldType))
                 {
-                    fieldType.CheckValue(ref value, copyBack: ref _ref, binder, culture, invokeAttr);
+                    fieldType.CheckValue(ref value, binder, culture, invokeAttr);
                 }
             }
             else if (!ReferenceEquals(value.GetType(), fieldType))
             {
-                fieldType.CheckValue(ref value, copyBack: ref _ref, binder, culture, invokeAttr);
+                fieldType.CheckValue(ref value, binder, culture, invokeAttr);
             }
 
             bool domainInitialized = false;