--- a/src/coreclr/System.Private.CoreLib/src/System/Runtime/RuntimeImports.CoreCLR.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Runtime/RuntimeImports.CoreCLR.cs
@@ -9,31 +9,9 @@ namespace System.Runtime
     internal static class RuntimeImports
     {
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern unsafe void RhRegisterForGCReporting(GCFrameRegistration* pRegistration);
+        internal static extern unsafe void RegisterForGCReporting(GCFrameRegistration* pRegistration);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        internal static extern unsafe void RhUnregisterForGCReporting(GCFrameRegistration* pRegistration);
-
-#pragma warning disable 0414, IDE0044
-        // Type that represents a managed view of the unmanaged GCFrame
-        // data structure in coreclr. The type layouts between the two should match.
-        internal unsafe struct GCFrameRegistration
-        {
-            private nuint m_reserved1;
-            private nuint m_reserved2;
-            private void** m_pObjRefs;
-            private uint m_numObjRefs;
-            private int m_MaybeInterior;
-
-            public GCFrameRegistration(void** allocation, uint elemCount, bool areByRefs = true)
-            {
-                m_reserved1 = 0;
-                m_reserved2 = 0;
-                m_pObjRefs = allocation;
-                m_numObjRefs = elemCount;
-                m_MaybeInterior = areByRefs ? 1 : 0;
-            }
-        }
-#pragma warning restore 0414, IDE0044
+        internal static extern unsafe void UnregisterForGCReporting(GCFrameRegistration* pRegistration);
     }
 }